fix: smooth remote movement (phase 1)
- server: broadcast position at 20Hz while moving (was ~1Hz -> stop-and-go)
- server: delta hash with full float precision (was toInt -> staircase)
- server: bufferDelayEnabled=false (inputs reach game loop immediately)
- server: use receive time for frames (client clock was double-offset)
- client: render buffer interpolation (fixed 80ms delay, no restarts)
- client: translate() no-op on remote components (single position owner)
- client: short 60ms idle snap, instant snap on real teleports
- client: shorter idle correction on local player
- ci: add analyze/test workflow

+++ game_server/src/game/state_tracker.dart
@@ -12,11 +12,13 @@ class MapStateTracker {
 
   /// Calculate hash for a ComponentStateModel (for change detection)
   int _calculateHash(ComponentStateModel state) {
-    // Hash based on position, direction, life, and action
-    // These are the fields that matter for visual updates
+    // Hash based on position, direction, life, and action.
+    // Uses full float precision (NOT toInt()): truncating to integers made
+    // sub-pixel movements undetectable, batching them into visible
+    // micro-teleports / corrections on remote clients.
     return Object.hash(
-      state.position.x.toInt(),
-      state.position.y.toInt(),
+      state.position.x,
+      state.position.y,
       state.direction,
       state.life,
       state.action,
